+++ src/main/java/com/myproject/idstudio/servlets/AuthnServlet.java
@@ -1,4 +1,4 @@
-package com.webapp.business_card.servlets;
+package com.myproject.idstudio.servlets;
 
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
@@ -10,8 +10,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.webapp.business_card.dao.CodeDao;
-import com.webapp.business_card.security.Encryptor;
+import com.myproject.idstudio.dao.CodeDao;
+import com.myproject.idstudio.security.Encryptor;
 
 @WebServlet("/authn")
 public class AuthnServlet extends HttpServlet {
