+++ src/main/webapp/WEB-INF/web.xml
@@ -7,31 +7,31 @@
 		<servlet-name>AdminServlet</servlet-name>
 		<display-name>AdminServlet</display-name>
 		<description></description>
-		<servlet-class>com.webapp.business_card.servlets.AdminServlet</servlet-class>
+		<servlet-class>com.myproject.idstudio.servlets.AdminServlet</servlet-class>
 	</servlet>
 	<servlet>
 		<servlet-name>AuthnServlet</servlet-name>
 		<display-name>AuthnServlet</display-name>
 		<description></description>
-		<servlet-class>com.webapp.business_card.servlets.AuthnServlet</servlet-class>
+		<servlet-class>com.myproject.idstudio.servlets.AuthnServlet</servlet-class>
 	</servlet>
 	<servlet>
 		<servlet-name>CallOrderServlet</servlet-name>
 		<display-name>CallOrderServlet</display-name>
 		<description></description>
-		<servlet-class>com.webapp.business_card.servlets.CallOrderServlet</servlet-class>
+		<servlet-class>com.myproject.idstudio.servlets.CallOrderServlet</servlet-class>
 	</servlet>
 	<servlet>
 		<servlet-name>PageServlet</servlet-name>
 		<display-name>PageServlet</display-name>
 		<description></description>
-		<servlet-class>com.webapp.business_card.servlets.PageServlet</servlet-class>
+		<servlet-class>com.myproject.idstudio.servlets.MainPageServlet</servlet-class>
 	</servlet>
 	<servlet>
 		<servlet-name>CodeChangeServlet</servlet-name>
 		<display-name>CodeChangeServlet</display-name>
 		<description></description>
-		<servlet-class>com.webapp.business_card.servlets.CodeChangeServlet</servlet-class>
+		<servlet-class>com.myproject.idstudio.servlets.CodeChangeServlet</servlet-class>
 	</servlet>
 
 </web-app>
